LiterateTest.jl

LiterateTest — Module

LiterateTest

Small utilities for writing testable documentation using Literate.jl.

LiterateTest.config — Function

LiterateTest.config(; overloads...) -> config::Dict{Symbol,Ans}

Construct a config dictionary that can be passed to Literate.markdown etc.

LiterateTest.config post-composes given preprocess function (default: identity) with LiterateTest.preprocess.

LiterateTest.preprocess — Function

LiterateTest.preprocess(code::AbstractString) -> code′::String

Remove testing related code from code.

See Testing LiterateTest.jl for examples of how it works.

Note

Currently, LiterateTest.preprocess is implemented as a plain line-based transformation. It may break with complex expressions.

LiterateTest.preprocess does the following transformations:

Remove ans = begin and end from

ans = begin
    # any number of lines with consistent indentation
end

and then de-indent the lines inside begin ... end.

Remove @test begin and end ... from

@test begin
    # any number of lines with consistent indentation
end # any content after `end` is ignored.

and then de-indent the lines inside begin ... end. Also remove global from assignments of the form global ... = ... if appears at the shallowest indentation level.

Remove the @testset block of the form

@testset ...
    ...
end

Extract $code from

@evaltest "$code" ...
    ...
end

and remove everything else.

LiterateTest.@evaltest — Macro

@evaltest "$code" begin $tests end
@evaltest raw"$code" begin $tests end

Evaluate code, assign it to the variable ans, and then run tests.

This is meant to be processed by LiterateTest.preprocess.

LiterateTest.AssertAsTest — Module

LiterateTest.AssertAsTest

A module that exports Test.@test as @assert.

An intended usage is to import @assert before including example file in test:

module MyTestExampels
using LiterateTest.AssertAsTest: @assert
include("../examples/file.jl")
end

This way, examples/file.jl can still use @assert for clean documentation while a better test failure messages are shown when running the tests.

Examples

julia> module MyTests
           using LiterateTest.AssertAsTest: @assert
           @assert 1 == 1
       end;